+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Unreleased
 ### Added
 - Added RECOMMENDATION to available import type to be pushed after feed is uploaded.
+- Added Check FTP connection functionality in the module configuration
 
 ## [v1.5.0] - 2020.03.06
 ### Changed